requirements.txt
torch>=1.3.1
torchvision>=0.4.2
pillow>=6.2.1